clicked area
increased the hitbox of the input on screen to character. so removed functions that rely on collider mouse click and made own if statements in update of touchable script. so now 1.5f outside character center is making it go to finger position so it is easier to touch when character is small(small collider).